Second phase of OMNIORB transport integration including GIOP transcoding in the omniORB transport driver. .gitignore: add time emitter data files to ignore list adapt/os/ocpios/linux/src/OcpiOsClientSocket.cxx: comment out debug adapt/os/ocpios/linux/src/OcpiOsSocket.cxx: ditto core/container/ctests/bin/include/test_utilities.h: remove obsolete header references core/container/ctests/bin/src/GPPContainerTest_main.cxx: ditto core/container/ctests/bin/src/test_advanced_port_mgmt_main.cxx: ditto core/container/ctests/bin/src/test_control_errors_main.cxx: ditto core/container/ctests/bin/src/test_optional_ports_main.cxx: ditto core/container/ctests/bin/src/test_shutdown_main.cxx: ditto core/container/ctests/bin/src/test_timeout_worker_main.cxx: ditto core/container/ctests/bin/src/test_transfer_roles_main.cxx: ditto core/container/ctests/bin/src/test_user_port_main.cxx: ditto core/container/ctests/bin/src/test_utilities.cxx: ditto core/container/ctests/bin/src/test_xml_worker_main.cxx: ditto core/container/ctests/bin/src/test_zero_copy_io_main.cxx: ditto core/container/rcc_container/include/OcpiRccPort.h: make ALL opcodes the uint8_t data type   - except in the RCC worker itself, which still uses RCCOrdinal core/container/rcc_container/src/OcpiRccPort.cxx: ditto core/container/rcc_container/src/OcpiRccMessagePort.cxx: ditto, plus   - standardize buffer method names so msg ports are same as others   - push some message port functionality down into the msg channel in preparation   - for the removal of message ports from rcc (so they are universal). core/dataplane/msg_driver_interface/include/DtMsgDriver.h: ditto, plus more const core/dataplane/msg_drivers/dds/src/OcpiDDSTransport.cxx: sync with interface   - changes in OcpiRccMessagePort.cxx core/dataplane/rdma_driver_interface/interfaces/include/DtHandshakeControl.h:   - add protocol info exchange offset/size fields core/dataplane/rdma_driver_interface/interfaces/src/DtTransferInternal.cxx:   - bug fixes (duplicate smbs), a memory leak core/dataplane/rdma_drivers/sockets/src/DtSocketsXfer.cxx: bug fixes, handle header fragments core/dataplane/rdma_tests/bin/src/ClientServerTest_main.cxx: update to also test protocol info exchange core/dataplane/transport/client_server/include/OcpiMessageCircuit.h: update for protocol info exchange   - also bring into sync with buffer interface consistency changes in OcpiRccPort.cxx   - also implement timeouts in message circuit calls core/dataplane/transport/client_server/include/OcpiMessageEndpoint.h: ditto core/dataplane/transport/client_server/src/OcpiMessageCircuit.cxx: ditto core/dataplane/transport/client_server/src/OcpiMessageEndpoint.cxx: ditto core/dataplane/transport/impl/include/OcpiCircuit.h: protocol info support core/dataplane/transport/impl/include/OcpiPort.h: protocol info and uniform opcode data types core/dataplane/transport/impl/include/OcpiTransport.h: ditto core/dataplane/transport/impl/src/OcpiCircuit.cxx: ditto core/dataplane/transport/impl/src/OcpiPort.cxx: ditto core/dataplane/transport/impl/src/OcpiTransport.cxx: ditto core/local/util/property/include/OcpiUtilDataTypes.h: support for noticing when      a data type is a sequence whose elements have fixed size, to support computing      the top level sequence's element count from message length in bytes core/local/util/property/include/OcpiUtilProtocol.h: ditto plus more const core/local/util/property/src/OcpiUtilDataTypes.cxx: ditto core/local/util/property/src/OcpiUtilProtocol.cxx: ditto plus support for parsing     protocols from xml strings, not as part of some larger file's xml tree jkmac_env_darwin_x86_64.sh: fix omni include install_omniorb.sh: fix leftover comment core/dataplane/rdma_smb/linux/src/OcpiPosixFileMappingServices.cxx: better debug core/dataplane/msg_drivers/corba/src/OcpiCORBATransport.cxx: openURL support for corbaloc

+++ core/dataplane/transport/impl/include/OcpiPort.h
@@ -129,7 +129,7 @@ namespace OCPI {
       /**********************************
        * Reteives the next available input buffer.
        *********************************/
-      BufferUserFacet* getNextFullInputBuffer(void *&data, uint32_t &length, uint32_t &opcode);
+      BufferUserFacet* getNextFullInputBuffer(void *&data, uint32_t &length, uint8_t &opcode);
       Buffer* getNextFullInputBuffer();
 
       /**********************************
@@ -199,7 +199,7 @@ namespace OCPI {
        * This method is used to send an input buffer thru an output port with Zero copy, 
        * if possible.
        *********************************/
-      void sendZcopyInputBuffer( Buffer* src_buf, unsigned int len, unsigned op );
+      void sendZcopyInputBuffer( Buffer* src_buf, unsigned int len, uint8_t op );
 
       /**********************************
        * This method causes the specified input buffer to be marked
@@ -211,7 +211,7 @@ namespace OCPI {
       /**********************************
        * Send an output buffer
        *********************************/
-      void sendOutputBuffer( BufferUserFacet* b, unsigned int length, unsigned int opcode );
+      void sendOutputBuffer( BufferUserFacet* b, unsigned int length, uint8_t opcode );
 
 
       // Advanced buffer management
